Port the platform service to be globally installed. (#5012)
Avoids having its code in extensions.
