+++ extensions/amp-analytics/0.1/test/test-crypto-impl.js
@@ -15,7 +15,7 @@
  */
 
 import {Crypto} from '../crypto-impl';
-import {Platform} from '../../../../src/platform';
+import {Platform} from '../../../../src/service/platform-impl';
 import * as lib from '../../../../third_party/closure-library/sha384-generated';
 import * as sinon from 'sinon';
