+++ extensions/amp-dynamic-css-classes/0.1/test/test-runtime-classes.js
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import {installPlatformService} from '../../../../src/service/platform-impl';
 import {installViewerService} from '../../../../src/service/viewer-impl';
 import {installVsyncService} from '../../../../src/service/vsync-impl';
 import {installDynamicClassesService} from '../amp-dynamic-css-classes';
@@ -55,6 +56,7 @@ describe('dynamic classes are inserted at runtime', () => {
   });
 
   function setup(embeded, userAgent, referrer) {
+    installPlatformService(mockWin);
     viewer = installViewerService(mockWin);
     const vsync = installVsyncService(mockWin);
