+++ src/platform.js
@@ -14,136 +14,14 @@
  * limitations under the License.
  */
 
-import {fromClass} from './service';
-
-
-/**
- * A helper class that provides information about device/OS/browser currently
- * running.
- */
-export class Platform {
-
-  /**
-   * @param {!Window} win
-   */
-  constructor(win) {
-    /** @const {!Navigator} */
-    this.navigator = /** @type {!Navigator} */ (win.navigator);
-  }
-
-  /**
-   * Whether the current platform an Android device.
-   * @return {boolean}
-   */
-  isAndroid() {
-    return /Android/i.test(this.navigator.userAgent);
-  }
-
-  /**
-   * Whether the current platform an iOS device.
-   * @return {boolean}
-   */
-  isIos() {
-    return /iPhone|iPad|iPod/i.test(this.navigator.userAgent);
-  }
-
-  /**
-   * Whether the current browser is Safari.
-   * @return {boolean}
-   */
-  isSafari() {
-    return /Safari/i.test(this.navigator.userAgent) && !this.isChrome() &&
-        !this.isEdge();
-  }
-
-  /**
-   * Whether the current browser is a Chrome browser.
-   * @return {boolean}
-   */
-  isChrome() {
-    // Also true for MS Edge :)
-    return /Chrome|CriOS/i.test(this.navigator.userAgent) && !this.isEdge();
-  }
-
-  /**
-   * Whether the current browser is a Firefox browser.
-   * @return {boolean}
-   */
-  isFirefox() {
-    return /Firefox/i.test(this.navigator.userAgent) && !this.isEdge();
-  }
-
-  /**
-   * Whether the current browser is a IE browser.
-   * @return {boolean}
-   */
-  isIe() {
-    return (/MSIE/i.test(this.navigator.userAgent) ||
-        /IEMobile/i.test(this.navigator.userAgent));
-  }
-
-  /**
-   * Whether the current browser is an Edge browser.
-   * @return {boolean}
-   */
-  isEdge() {
-    return /Edge/i.test(this.navigator.userAgent);
-  }
-
-  /**
-   * Whether the current browser is based on the WebKit engine.
-   * @return {boolean}
-   */
-  isWebKit() {
-    return /WebKit/i.test(this.navigator.userAgent) && !this.isEdge();
-  }
-
-  /**
-   * Returns the major version of the browser.
-   * @return {number}
-   */
-  getMajorVersion() {
-    if (this.isSafari()) {
-      return this.evalMajorVersion_(/\sVersion\/(\d+)/, 1);
-    }
-    if (this.isChrome()) {
-      return this.evalMajorVersion_(/\Chrome\/(\d+)/, 1);
-    }
-    if (this.isFirefox()) {
-      return this.evalMajorVersion_(/\Firefox\/(\d+)/, 1);
-    }
-    if (this.isIe()) {
-      return this.evalMajorVersion_(/\MSIE\s(\d+)/, 1);
-    }
-    if (this.isEdge()) {
-      return this.evalMajorVersion_(/\Edge\/(\d+)/, 1);
-    }
-    return 0;
-  }
-
-  /**
-   * @param {!RegExp} expr
-   * @param {number} index The index in the result that's interpreted as the
-   *   major version (integer).
-   * @return {number}
-   */
-  evalMajorVersion_(expr, index) {
-    if (!this.navigator.userAgent) {
-      return 0;
-    }
-    const res = this.navigator.userAgent.match(expr);
-    if (!res || index >= res.length) {
-      return 0;
-    }
-    return parseInt(res[index], 10);
-  }
-};
+import {getExistingServiceForWindow} from './service';
 
 
 /**
  * @param {!Window} window
- * @return {!Platform}
+ * @return {!./service/platform-impl.Platform}
  */
 export function platformFor(window) {
-  return fromClass(window, 'platform', Platform);
+  return /** @type {!./service/platform-impl.Platform} */ (
+      getExistingServiceForWindow(window, 'platform'));
 };
