+++ src/service/ios-scrollfreeze-bug.js
@@ -31,7 +31,7 @@ import {vsyncFor} from '../vsync';
  * See #3481 for more details.
  *
  * @param {!Window} win
- * @param {!../platform.Platform=} opt_platform
+ * @param {!../service/platform-impl.Platform=} opt_platform
  * @param {!./viewer-impl.Viewer=} opt_viewer
  * @param {!./vsync-impl.Vsync=} opt_vsync
  * @return {?Promise}
