+++ test/functional/test-fixed-layer.js
@@ -15,6 +15,7 @@
  */
 
 import {FixedLayer} from '../../src/service/fixed-layer';
+import {installPlatformService} from '../../src/service/platform-impl';
 import * as sinon from 'sinon';
 
 
@@ -133,6 +134,7 @@ describe('FixedLayer', () => {
       documentElement: docElem,
       body: docBody,
     };
+    installPlatformService(documentApi.defaultView);
 
     vsyncTasks = [];
     vsyncApi = {
